# Introduce cycle-based gas cost model.