--- a/execution_engine/Cargo.toml
+++ b/execution_engine/Cargo.toml
@@ -49,6 +49,9 @@ tempfile = "3.4.0"
 thiserror = "1.0.18"
 tracing = "0.1.18"
 uint = "0.9.0"
+clap = { version = "4.5.21", features = ["derive"] }
+toml = "0.8.19"
+wat = "1.220.0"
 
 [dev-dependencies]
 assert_matches = "1.3.0"