--- a/resources/local/chainspec.toml.in
+++ b/resources/local/chainspec.toml.in
@@ -204,9 +204,9 @@ vm_casper_v2 = false
 # [3] -> Transaction gas limit size in bytes for a given transaction in a certain lane
 # [4] -> The maximum number of transactions the lane can contain
 native_mint_lane = [0, 1024, 1024, 65_000_000_000, 650]
-native_auction_lane = [1, 2048, 2048, 362_500_000_000, 145]
-install_upgrade_lane = [2, 1_048_576, 2048, 1_000_000_000_000, 1]
-wasm_lanes = [[3, 344_064, 1024, 500_000_000_000, 3], [4, 172_032, 1024, 50_000_000_000, 7], [5, 12_288, 512, 1_500_000_000, 15]]
+native_auction_lane = [1, 2048, 2048, 2_500_000_000, 145]
+install_upgrade_lane = [2, 1_048_576, 2048, 100_000_000_000, 1]
+wasm_lanes = [[3, 344_064, 1024, 100_000_000_000, 3], [4, 172_032, 1024, 50_000_000_000, 7], [5, 12_288, 512, 1_500_000_000, 15]]
 
 [transactions.deploy]
 # The maximum number of Motes allowed to be spent during payment.  0 means unlimited.